requirements.txt
tensorflow
numpy
pandas
scikit-learn
dvc
ipykernel
mlflow
notebook
matplotlib
seaborn
python-dotenv
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
streamlit
kagglehub
ipykernel
-e .